poetry should now work
