+++ apps/client/src/routes/poetry/[id]/__layout.svelte
@@ -1,4 +1,6 @@
-<script lang="ts" context="module">
+<script context="module" lang="ts">
+    /// This layout file is required so that works are fetched before anything else gets touched.
+
     import { get } from 'svelte/store';
     import type { Load } from '@sveltejs/kit';
     import { CommentKind } from '$lib/models/comments';
@@ -15,12 +17,12 @@
             get(session).currProfile ? get(session).currProfile._id : undefined,
         );
 
-        setContent(poetry.content, poetry.content.sections, poetry.ratings, poetry.libraryDoc);
+        setContent(poetry);
 
         return {
             props: {
-                content: poetry.content,
-                comments: await getPage(poetry.content._id, CommentKind.ContentComment, page),
+                content: poetry,
+                comments: await getPage(poetry._id, CommentKind.ContentComment, page),
             },
         };
     };
